+++ R/to_label.R
@@ -211,7 +211,7 @@ to_label_helper <- function(x, add.non.labelled, prefix, var.label, drop.na, dro
     }
   }
   # drop unused levels?
-  if (drop.levels) x <- droplevels(x)
+  if (drop.levels && is.factor(x)) x <- droplevels(x)
 
   # set back variable labels
   if (!is.null(var_lab)) x <- suppressWarnings(sjlabelled::set_label(x, label = var_lab))